+++ cookbooks/mylab_firewall/metadata.rb
@@ -1,23 +1,22 @@
-name 'maxlab_firewall'
-maintainer 'Maxwell Spangler'
-maintainer_email 'maxcode@maxwellspangler.com'
+name 'mylab_firewall'
+maintainer 'The Authors'
+maintainer_email 'you@example.com'
 license 'All Rights Reserved'
-description 'Installs/Configures maxlab_firewall'
-long_description 'Installs/Configures maxlab_firewall'
-version '1.2.0'
-chef_version '>= 13.0'
+description 'Installs/Configures mylab_firewall'
+version '1.0.0'
+chef_version '>= 14.0'
 
 # The `issues_url` points to the location where issues for this cookbook are
 # tracked.  A `View Issues` link will be displayed on this cookbook's page when
 # uploaded to a Supermarket.
 #
-# issues_url 'https://github.com/<insert_org_here>/maxlab_firewall/issues'
+# issues_url 'https://github.com/<insert_org_here>/mylab_firewall/issues'
 
 # The `source_url` points to the development repository for this cookbook.  A
 # `View Source` link will be displayed on this cookbook's page when uploaded to
 # a Supermarket.
 #
-# source_url 'https://github.com/<insert_org_here>/maxlab_firewall'
+# source_url 'https://github.com/<insert_org_here>/mylab_firewall'
 
 supports 'redhat', '>= 7.0'
 supports 'centos', '>= 7.0'